% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{betaStats}
\alias{betaStats}
\title{Calculates the mean and variance of a Beta distribution.}
\usage{
betaStats(a, b)
}
\arguments{
\item{a}{first Beta shape parameter}

\item{b}{second Beta shape parameter}
}
\value{
named vector with values for mean and variance
}
\description{
Given the parameters \code{(a, b)} of a Beta distribution, this
function calculates the mean and variance as:
\code{mean = a / (a + b)} and
\code{variance = (ab) / ((a+b+1)(a+b)^2)}
}

