% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/burr_distribution.R
\name{dburr}
\alias{dburr}
\alias{pburr}
\alias{rburr}
\title{The Burr XII distribution}
\usage{
dburr(x, c, k, scale = 1)

pburr(q, c, k, scale = 1)

rburr(n, c, k, scale = 1)
}
\arguments{
\item{x,}{q vector of quantiles (> 0)}

\item{c,}{k shape parameters (> 0)}

\item{scale}{scale parameter (> 0)}

\item{n}{number of observations or, if length(n) > 1, the length is
  taken to be the number required}
}
\description{
Density, distribution function and random generation for the Burr Type XII distribution
(commonly called more simply the Burr distribution).
}
\details{
The Burr type XII distribution with shape parameters \eqn{c, k} and scale parameter \eqn{s}
has density:

\deqn{
  (ck/s)[ (x/s)^(c - 1) ] / [ (1 + (x/s)^c)^(k + 1) ]
}

for \eqn{x > 0} and \eqn{c,k,s > 0}.

The cumulative probability function is:

\deqn{
  1 - ( 1 + (x/s)^c )^(-k)
}

The mean is given by \eqn{\mu(1)} and the variance by \eqn{-\mu(1)^2 + \mu(2)} where:

\deqn{ \mu(i) = (ks^i)Beta( (ck - i)/c, (c + i)/c ) }

The median is: \eqn{ s(2^(1/k) - 1)^(1/c) }

The mode is: \eqn{ s( (c - 1)/(ck + 1) )^(1/c) }

The Burr XII distribution is also known as the generalized log-logistic distribution.

When shape parameter \eqn{c = 1} the Burr XII is the Pareto II distribution.
When shape parameter \eqn{k = 1} it is a special case of the Fisk distribution.
}

