% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertParamList.R
\name{convertParamList}
\alias{convertParamList}
\title{Convert an old-format species parameters object to a SpeciesParams object}
\usage{
convertParamList(oldparams)
}
\arguments{
\item{oldparams}{named list of parameter values}
}
\value{
a new \code{SpeciesParams} object
}
\description{
Converts an old-format species parameters object, as used by an earlier version of
the simulation function, into a \linkS4class{SpeciesParams} object.
}

