% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeciesParams-class.R
\docType{methods}
\name{getFunctionOrElse,SpeciesParams-method}
\alias{getFunctionOrElse,SpeciesParams-method}
\title{Gets an optional function if present, or a fallback function object.}
\usage{
\S4method{getFunctionOrElse}{SpeciesParams}(object, fnName, fallback)
}
\arguments{
\item{object}{\code{SpeciesParam} object}

\item{fnName}{name of the function slot (partial matching is used)}

\item{fallback}{fallback function object}
}
\value{
the optional function if present, or the fallback
}
\description{
Gets an optional function if present, or a fallback function object.
}
\examples{
\dontrun{
params <- new("SpeciesParams", "test")

## Ask for the recruit_canopy_fn object. Since we haven't set one
## we will get the fallback (identity function in this example)
fn <- getFunctionOrElse(params, "recruit_canopy_fn", identity)
}
}

