% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpeciesParams-class.R
\docType{methods}
\name{isComplete,list-method}
\alias{isComplete,list-method}
\title{Check if each \code{SpeciesParams} object in a list is ready for use.}
\usage{
\S4method{isComplete}{list}(object)
}
\arguments{
\item{object}{a list of \code{SpeciesParams} objects to check}
}
\value{
a character vector of error messages for invalid or incomplete
  elements; otherwise an empty vector if all elements are valid
}
\description{
Checks that all required elements of the given \code{SpeciesParams} object
have valid values.
}

