% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpeciesParams-class.R
\docType{methods}
\name{isComplete,SpeciesParams-method}
\alias{isComplete,SpeciesParams-method}
\title{Check if a \code{SpeciesParams} object is ready for use.}
\usage{
\S4method{isComplete}{SpeciesParams}(object)
}
\arguments{
\item{object}{the object to check}
}
\value{
a character vector of error messages for invalid or incomplete
  elements; otherwise an empty vector if all elements are valid
}
\description{
Checks that all required elements of the given \code{SpeciesParams} object
have valid values.
}
\examples{
\dontrun{
params <- new("SpeciesParams", name="Tree")

## put data for each element in slot
params@max_height <- 30
params@height_yr1 <- 0.1
params@tree_desc <- my.tree.desc.matrix

## plus the rest, then...
errs <- isComplete(params)
if (length(errs) > 0)
  cat(errs, sep="\\n")
else
  ## params valid to use
}
}

