% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resampleSerial.R
\name{resampleSerial}
\alias{resampleSerial}
\title{Resample a vector of serially-autocorrelated values.}
\usage{
resampleSerial(x, length.out, win.size = c(10, 20), wrap = FALSE,
  na.action = c("allow", "omit", "reject"))
}
\arguments{
\item{x}{Vector to resample}

\item{length.out}{Length of the output sequence}

\item{win.size}{Either a vector of two values giving the lower and
  upper bounds of resampling windows, or a single value for constant
  window size}

\item{wrap}{If TRUE, the input time series is treated as circular
  (see Details).}

\item{na.action}{Controls how sampling deals with NA values in the
  input vector.
  \describe{
    \item{"allow"}{No special action: NAs are sampled as any other value.}
    \item{"omit"}{If any NAs occur in a window they are simply omitted.}
    \item{"reject"}{If any NAs occur in a window, that window is
      rejected and another chosen.}
  }}
}
\description{
This function performs resampling with replacement on an input vector of
(possibly) serially autocorrelated values. Values within randomly placed
windows are collected and concatenated to form the output vector. Window
sizes can be random within specified bounds, or constant. The aim of the
function is to produce a resampled version of the input data while
maintaining the pattern of serial autocorrelation as much as possible. To
achieve this, you will probably want to experiment with window sizes, using
ACF plots or similar to check results.
}
\details{
By default, the input vector is treated as non-circular. If a randomly placed
window extends beyond the right-hand edge of the input vector, it will be
rejected and a new window created. Alternatively, setting the \code{wrap}
argument to TRUE allows the input vector to be treated as circular. In this
case, a window which extends beyond the right-hand edge of the input vector
will wrap around to the left-hand edge.
}

