% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdb_utility_functions.R
\name{tmdbClose}
\alias{tmdbClose}
\title{Close a database connection.}
\usage{
tmdbClose(tmdb, deleteDB = FALSE)
}
\arguments{
\item{tmdb}{an open database connection}

\item{deleteDB}{whether to delete the database file from disk}
}
\description{
Closes the given database connection and optionally
deletes the associated database file.
}
\examples{
\dontrun{
con <- tmdbOpen("somedir/results.db")

# do interesting things, then...

tmdbClose(con)
}
}
\seealso{
\code{\link{tmdbOpen}}
}

