% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmdbGetExtant.R
\name{tmdbGetExtant}
\alias{tmdbGetExtant}
\title{Summarize species extant status}
\usage{
tmdbGetExtant(tmdb, paramSetID, time)
}
\arguments{
\item{tmdb}{an open database connection}

\item{paramSetID}{identifier of the parameter set to query}

\item{time}{the simulation time to query}
}
\value{
A list with the following elements:
  paramSetID (parameter set identifier);
  numRuns (number of runs found for the parameter set);
  time (query time);
  data (a `data.frame` with columns spID, name, runs.extant, prop)
}
\description{
For a given parameter set, identified by its integer ID, finds the number
of runs in which each modelled species was extant at a specified time.
}

