% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmdbGetSplat.R
\name{tmdbGetSplat}
\alias{tmdbGetSplat}
\title{Gets the stand data for a specified run and year (a 'splat').}
\usage{
tmdbGetSplat(tmdb, runid = 1, time)
}
\arguments{
\item{tmdb}{an open database connection}

\item{runid}{identifier(s) (integer) of the simulation run(s) to query}

\item{time}{simulation year (integer)}
}
\value{
`data.frame` with columns: CanopyRadius, N, SpeciesID
}
\description{
Gets the stand data for a specified run and year (a 'splat').
}
\examples{
\dontrun{
con <- tmdbOpen("results.db")

# query a year across several runs
dat <- tmdbGetSplat(con, runid=1:3, time=100)

# close the database connection
tmdbClose(con)
}
}

