% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmdb_utility_functions.R
\name{tmdbSave}
\alias{tmdbSave}
\title{Save a database to a new file.}
\usage{
tmdbSave(tmdb, dest, reconnect = FALSE)
}
\arguments{
\item{tmdb}{an open database connection}

\item{dest}{destination path}

\item{reconnect}{if TRUE a new connection to the destination
  database file is returned}
}
\value{
Invisibly returns a new connection (if reconnect is TRUE)
  or NULL otherwise.
}
\description{
Copies a database from its current location (possibly a temporary file)
to the given destination file. After copying, the connection is closed
and the original file deleted.
}

