% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdbSummary.R
\name{tmdbSummary}
\alias{tmdbSummary}
\title{Summarize a simulation output database.}
\usage{
tmdbSummary(tmdb)
}
\arguments{
\item{tmdb}{an open database connection}
}
\value{
A data.frame with columns RunID and TimeLength.
}
\description{
Queries the output database for the ID value and time length of each
run contained within it.
}

