% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmDriver.R
\name{tmDriver}
\alias{tmDriver}
\title{Driver function}
\usage{
tmDriver(funs, nmajor = 1, nminor = 1)
}
\arguments{
\item{funs}{A named list of generating functions, where the names correspond
  to \code{\link{tmRun}} argument names. Each function takes two integer
  parameters (major and minor run number) and returns an object appropriate
  for the particular argument (e.g. a generating function for the \code{'rain'})
  argument would return a vector of annual rainfall values.}
}
\description{
Generating functions take two parameters: major and minor run number. Each
function is free to interpret these in any way, or ignore them completely.
For example, in a study of rainfall effects, major number may relate to an
overall average rainfall value while the minor number is simply the replicate
index within each rainfall average.
}

