% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmRun.R
\name{tmRun}
\alias{tmRun}
\title{Simulates dynamics of a woodland stand.}
\usage{
tmRun(Spp, initial.cohorts, rain, scheduled.fires = NULL, fire.func = NULL,
  fire.canopy.func = NULL, fire.early.prob = 0, overlap.matrix = 0,
  ov.gen = 0, recruitment = TRUE, thinning = NULL, special = NULL,
  seedling.survival = NULL, session.settings = NULL, database = NULL)
}
\arguments{
\item{Spp}{either a single object or a \code{list} of objects
  of class \code{\linkS4class{SpeciesParams}} defining one or more species}

\item{initial.cohorts}{numeric matrix with 4 columns:
  species index, age, height, number of trees.
  For simulations with a single species the species index should be 1, while
  for multiple species the index value is the position in the list
  passed to the \code{Spp} argument.}

\item{rain}{numeric vector of annual rainfall values. The length of this
  vector determines the number of years in the simulation.}

\item{scheduled.fires}{numeric vector of fire intensity values. The scale is
  user-defined. If provided, the \code{fire.func} argument must also be
  present.}

\item{fire.func}{a function taking two numeric arguments (flammable stand
  proportion and incoming fire intensity) and returning a single numeric
  value for realized fire intensity}

\item{fire.canopy.func}{a function taking a single numeric argument for
  summed core canopy area, and returning a (possibly) transformed area value.
  If not provided, untransformed canopy area will be used.}

\item{overlap.matrix}{WRITE ME !}

\item{ov.gen}{WRITE ME !}

\item{recruitment}{WRITE ME !}

\item{thinning}{WRITE ME !}

\item{special}{WRITE ME !}

\item{seedling.survival}{WRITE ME !}

\item{session.settings}{WRITE ME !}

\item{database}{an optional existing SQLite database connection to write results
  and metadata to. If not provided, a new database will be created.}

\item{the}{probability that fire will occur before recruitment in any given
  year. The default value (0) specifies that fires always occur later than
  recruitment.}
}
\value{
a connection to the output SQLite database. If the \code{database}
  argument was NULL (default) then this will point to a temporary file which
  can be saved using the \code{\link{tmdbSave}} function. If an open
  connection to an existing database was provided via the \code{database}
  argument, then this same connection will be returned.
}
\description{
\code{tmRun} simulates the dynamics of cohorts of one or more tree species
in a woodland stand. For each cohort the simulation tracks annual values of
age, number of trees, height, plus a set of morphometrics derived from height
such as DBH and canopy radius. The location of individual trees is not
explicitly modelled. The stand environment is defined by input vectors for
annual rainfall and fire occurrence / intensity. The simulation period is
equal to the length of the rainfall vector.
}
\details{
The order of events in each year of the simulation is:
\enumerate{
  \item Optional special actions (coppice ON)
  \item Optional thinning of trees
  \item Growth of trees (uniform within cohort)
  \item Survival (adjustment of numbers within cohorts)
  \item Recruitment (addition of new cohorts)
  \item Fire (possible mortality and reduction of tree height)
  \item Report cohort statistics
}
}

