% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/woodland.R
\docType{package}
\name{woodland}
\alias{woodland}
\alias{woodland-package}
\title{woodland: A package for woodland stand simulation.}
\description{
A cohort-level simulation model of woodland tree dynamics.
}
\references{
Karen A. Ross, Ian D. Lunt, Ross A. Bradstock, Michael Bedward, Murray V. Ellis.
  Did historical tree removal promotewoody plant encroachment in Australian woodlands ?
  Journal of Vegetation Science 23 (2012) 304-312.

Karen A. Ross, Michael Bedward, Murray V. Ellis, Andrew Deane,
  Christopher C. Simpson, Ross A. Bradstock.
  Modelling the dynamics of white cypress pine Callitris glaucophylla woodlands
  in inland south-eastern Australia.
  Ecological Modelling 211 (2008) 11-24.
}

