% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdb_plot_functions.R
\name{tmdbPlot}
\alias{tmdbPlot}
\title{Plots yearly summed value (across cohorts by species) for a chosen variable.}
\usage{
tmdbPlot(tmdb, val = c("mergedarea", "basalarea", "corearea", "resourceuse",
  "n"), runid = 1, show.legend = TRUE, plot.total = TRUE)
}
\arguments{
\item{tmdb}{an open connection to a \code{\link{tmRun}} output database}

\item{val}{name (character) of the variable to plot. The values plotted
  will be the sum of individual cohort values for each year.}

\item{runid}{a vector of one or more run ID values, each of which will
  be used to retrieve data for a separate plot}

\item{show.legend}{if TRUE, legend is added to each plot}

\item{plot.total}{if TRUE, and there are two or more species in the results,
  plot the combined value in addition to that for each species}
}
\description{
Plots yearly summed value (across cohorts by species) for a chosen variable.
}

