% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmdb_plot_functions.R
\name{tmdbPlot}
\alias{tmdbPlot}
\title{Plot simulation outputs.}
\usage{
tmdbPlot(tmdb, variable = c("n", "mergedarea", "basalarea", "corearea",
  "resourceuse"), run = 1, show.legend = TRUE, plot.total = TRUE,
  plot.fire = TRUE, extra.sql = "")
}
\arguments{
\item{tmdb}{an open connection to a \code{\link{tmRun}} output database}

\item{variable}{The name (character) of the variable to plot: \code{"n",
  "mergedarea", "basalarea", "corearea", "resourceuse"} (may be abbreviated).
  The values plotted will be the sum of individual cohort values for each
  year.}

\item{run}{A vector of one or more run ID values (default = 1), each of which
  will be used to retrieve data for a separate plot}

\item{show.legend}{If TRUE (default), legend is added to each plot}

\item{plot.total}{If TRUE (default), and there are two or more species in the
  results, plot the combined value in addition to that for each species}

\item{plot.fire}{If TRUE (default), adds fire occurrences as a rug plot.}

\item{extra.sql}{Additional conditions to apply when querying the database
  for plot data. See example.}
}
\value{
A ggplot object.
}
\description{
Plots yearly value, summed across cohorts by species, for a chosen variable.
}
\examples{
\dontrun{
## Plot the number of trees of each species in runs 1 to 4
tmdbPlot(con, plot="n", run=1:4)

## Same plot but for subset of trees with height 5-10m
tmdbPlot(con, plot="n", run=1:4, extra.sql="height >= 5 and height <= 10")
}
}

